#2 - Add: Created controller for hot topics
