--- src/main/resources/application.properties
@@ -1,5 +1,0 @@
-spring.datasource.url=jdbc:h2:file:/opt/jpadb
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=mypass
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
